+++ src/Models/Tenant.php
@@ -2,7 +2,10 @@
 
 namespace Placetopay\Cerberus\Models;
 
+use Database\Factories\TenantFactory;
+use Illuminate\Database\Eloquent\Factories\Factory;
 use Placetopay\Cerberus\Scopes\AppScope;
+use Spatie\Multitenancy\Contracts\IsTenant;
 use Spatie\Multitenancy\Models\Tenant as TenantSpatie;
 
 /**
@@ -12,7 +15,7 @@ use Spatie\Multitenancy\Models\Tenant as TenantSpatie;
  * @property string domain
  * @property array config
  */
-class Tenant extends TenantSpatie
+class Tenant extends TenantSpatie implements IsTenant
 {
     protected $fillable = ['config'];
 
@@ -67,5 +70,10 @@ class Tenant extends TenantSpatie
     private function shortLocale($locale): string
     {
         return strtolower(substr($locale, 0, 2));
+    }
+
+    protected static function newFactory(): Factory
+    {
+        return TenantFactory::new();
     }
 }