+++ src/TenantFinder/DomainTenantFinder.php
@@ -4,16 +4,13 @@ namespace Placetopay\Cerberus\TenantFinder;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Cache;
+use Spatie\Multitenancy\Contracts\IsTenant;
 use Spatie\Multitenancy\Landlord;
-use Spatie\Multitenancy\Models\Concerns\UsesTenantModel;
-use Spatie\Multitenancy\Models\Tenant;
 use Spatie\Multitenancy\TenantFinder\TenantFinder;
 
 class DomainTenantFinder extends TenantFinder
 {
-    use UsesTenantModel;
-
-    public function findForRequest(Request $request): ?Tenant
+    public function findForRequest(Request $request): ?IsTenant
     {
         $domain = $request->getHost().str_replace('/index.php', '', $request->getBaseUrl());
         $vanityUrl = $_ENV['APP_VANITY_URL'] ?? '';
@@ -25,11 +22,11 @@ class DomainTenantFinder extends TenantFinder
         return $this->getTenant($domain);
     }
 
-    public function getTenant($domain)
+    public function getTenant($domain): ?IsTenant
     {
         return Landlord::execute(function () use ($domain) {
-            return Cache::rememberForever("tenant_{$domain}", function () use ($domain) {
-                return $this->getTenantModel()::query()->whereDomain($domain)->first();
+            return Cache::rememberForever("tenant_$domain", function () use ($domain) {
+                return app(IsTenant::class)::query()->whereDomain($domain)->first();
             });
         });
     }