+++ tests/Feature/Commands/TenantsListCommandTest.php
@@ -2,6 +2,7 @@
 
 namespace Placetopay\Cerberus\Tests\Feature\Commands;
 
+use PHPUnit\Framework\Attributes\Test;
 use Placetopay\Cerberus\Models\Tenant;
 use Placetopay\Cerberus\Tests\TestCase;
 
@@ -11,14 +12,14 @@ class TenantsListCommandTest extends TestCase
     {
         parent::setUp();
 
-        factory(Tenant::class)->create([
+        Tenant::factory()->create([
             'app' => config('multitenancy.identifier'),
             'name' => 'tenant_1',
             'domain' => 'co.domain.com',
             'config' => $this->getConfigStructure('laravel_mt_tenant_1'),
         ]);
 
-        factory(Tenant::class)->create([
+        Tenant::factory()->create([
             'app' => 'other',
             'name' => 'tenant_2',
             'domain' => 'pr.domain.com',
@@ -26,7 +27,7 @@ class TenantsListCommandTest extends TestCase
         ]);
     }
 
-    /** @test */
+    #[Test]
     public function it_can_list_the_tenencies_of_the_app()
     {
         $this